requirements.txt
# Core RAG
transformers==4.38.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
scipy==1.11.4
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.2

# LangChain + Vector Store (FAISS)
langchain==0.1.16
faiss-cpu==1.7.4
huggingface-hub==0.22.2

# Document loaders
pypdf==3.17.4
python-dotenv==1.0.1
langchain-community==0.0.5  

# Google Gemini
google-generativeai==0.3.2  
langchain-google-genai==0.0.3  

# Compatibility for Mac M1
protobuf==4.25.3        
grpcio==1.62.1         
